ChatRoom issues fixed.

+++ composeApp/src/commonMain/kotlin/com/srisu/srisu/core/data/apiservice/profile/ProfileApiService.kt
@@ -44,7 +44,7 @@ class ProfileApiService(private val httpClient: HttpClient) {
     }
 
     suspend fun sendUpdateProfileRequest(
-        userId: Int?,
+        userId: Long?,
         profileUpdateDTO: ProfileUpdateDTO,
         mediaFile: MediaFile?,
         gallery: List<MediaFile?>?
